Task updated

+++ 8-kyu/sum-of-differences-in-array/Reader.java
@@ -3,13 +3,13 @@ import java.util.Scanner;
 
 public class Reader {
 
-    public static double[] readArray() {
+    public static int[] readArray() {
 
         Scanner scanner = new Scanner(System.in);
 
         String nextLine = scanner.nextLine();
 
-        return nextLine.equals("[]") ? new double[0] :
+        return nextLine.equals("[]") ? new int[0] :
 
                 Arrays.stream(nextLine
                         .replace(",", " ")
@@ -17,7 +17,7 @@ public class Reader {
                         .replace("]", " ")
                         .trim()
                         .split("\\s+"))
-                        .mapToDouble(Double::parseDouble)
+                        .mapToInt(Integer::parseInt)
                         .toArray();
 
     }
